moving configurator in dedicated module moving other objects in utils
